+++ lib/src/mutation_test.dart
@@ -139,6 +139,7 @@ class MutationTest {
   /// Count all mutations done in all input files
   Future<void> _countAll() async {
     var totalCount = 0;
+    var fileCount = 0;
     for (final file in inputs) {
       var data = _createMutationData(file, outputPath, false, dry, format,
           ruleFiles: ruleFiles, addBuiltin: builtinRules);
@@ -149,6 +150,7 @@ class MutationTest {
 
         var count = await countMutations(data);
         totalCount += count;
+        fileCount += 1;
       }
     }
     if (inputs.isEmpty) {
@@ -163,11 +165,10 @@ class MutationTest {
 
         var count = await countMutations(data);
         totalCount += count;
+        fileCount += 1;
       }
     }
-    if (verbose) {
-      print('Performing $totalCount mutations!');
-    }
+    print('Found $totalCount mutations in $fileCount source files!');
     bar.mutationCount = totalCount;
   }